Refactor tests to share logic
