+++ cypress/e2e/new-suggestion/instruments/new-suggestion-instruments.cy.ts
@@ -1,22 +1,15 @@
+import { addInstrumentItem } from "./new-suggestion-instruments"
+import { fillSongInformationSuccessfully } from "../song-information/new-suggestion-song-information"
+
 const toInstrumentsProgressButton = "new-suggestion-progress-bar-instruments"
 const instrumentsArea = "area-instruments"
 const toReviewButton = "to-review-button"
 const searchInstrumentInput = "search-instrument-input"
-
 const instrumentSearchList = "instrument-search-list"
 const instrumentEditList = "instrument-edit-list"
 const deleteButton = "delete-button"
 const instrumentSearchCloseButton = "instrument-search-close-button"
 const descriptionInput = "description-input"
-const inputTitle = "input-title"
-const inputArtist = "input-artist"
-const inputLink = "input-link"
-const inputMotivation = "input-motivation"
-
-const addInstrumentItem = () => {
-  cy.data(searchInstrumentInput).type("a")
-  cy.data(instrumentSearchList).first().click()
-}
 
 describe("when creating new instrument suggestions for a suggestion", () => {
   beforeEach(() => {
@@ -24,10 +17,7 @@ describe("when creating new instrument suggestions for a suggestion", () => {
     cy.visit("/suggestions/new")
     // Wait so content can render properly and set up submit events
     cy.wait(500)
-    cy.data(inputTitle).type("Hello")
-    cy.data(inputArtist).type("Hello")
-    cy.data(inputLink).type("www.hello.com")
-    cy.data(inputMotivation).type("Hello")
+    fillSongInformationSuccessfully()
     cy.data(toInstrumentsProgressButton).click()
   })
