--- cypress/fixtures/state-filled-in-instruments.json
@@ -1,19 +1,0 @@
-{
-  "newSuggestion": {
-    "activeArea": "song-information",
-    "suggestion": {
-      "title": "Let It Be",
-      "artist": ["The Beatles"],
-      "link": null,
-      "motivation": "We have already sung it once while just playing randomly and it was pretty fun so thought it would be nice to add it to the repertoire.",
-      "instruments": [
-        {
-          "id": "f5ccaa5d-c601-40fb-8604-5fc7485f8528",
-          "instrument_name": "Acoustic Guitar",
-          "image_source": "guitar",
-          "description": "note"
-        }
-      ]
-    }
-  }
-}
