--- cypress/fixtures/state-filled-in-song-information.json
@@ -1,12 +1,0 @@
-{
-  "newSuggestion": {
-    "activeArea": "song-information",
-    "suggestion": {
-      "title": "Let It Be",
-      "artist": ["The Beatles"],
-      "link": null,
-      "motivation": "We have already sung it once while just playing randomly and it was pretty fun so thought it would be nice to add it to the repertoire.",
-      "instruments": []
-    }
-  }
-}
